+++ main/coins/nano/menus/address_text.c
@@ -24,10 +24,10 @@
 #include "nano_lib.h"
 
 #include "menu8g2.h"
-#include "../../../vault.h"
 #include "submenus.h"
 #include "../../../globals.h"
-#include "../../../statusbar.h"
+#include "../../../vault.h"
+#include "../../../gui/statusbar.h"
 
 static const char* TAG = "address_text";
 static const char TITLE[] = "Nano Address";
