+++ main/component.mk
@@ -1,6 +1,11 @@
 # Add coin directories and subdirectoreis below
 COMPONENT_SRCDIRS := . \
+	gui \
+	gui/entry \
+	gui/menus \
+	hal \
+	radio \
+	syscore \
 	coins/nano \
-	coins/nano/menus \
-	menus \
-	syscore
+	coins/nano/menus
+
