+++ main/gui/qr.c
