+++ main/gui/qr.h
